+++ secrets.go
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2025 The templig contributors.
+// SPDX-FileCopyrightText: 2026 The templig contributors.
 // SPDX-License-Identifier: MPL-2.0
 
 package templig
@@ -24,13 +24,21 @@ type secretWorkItem struct {
 	secret bool
 }
 
-// HideSecrets hides secrets in the given YAML node structure. Secrets are identified using the [SecretRE].
+// HideSecrets hides secrets in the given YAML node structure.
+// Secrets are identified using the given `secretRE` parameter. If that parameter is nil, [SecretDefaultRE] is used
+// instead to construct a new regexp to prevent silent complete failures to hide secrets.
 // Depending on the parameter `hideStructure`, the structure of the secret is hidden too (`true`) or visible (`false`).
-func HideSecrets(node *yaml.Node, hideStructure bool) {
+func HideSecrets(node *yaml.Node, hideStructure bool, secretRE *regexp.Regexp) {
 	// initialWorkQueueDepth is an assumption about the maximum depth of the YAML document structure. It will not limit
 	// the real depth, but should, for average use cases of templig, be enough.
 	const initialWorkQueueDepth = 20
 	const queueCompactionFrequency = 100
+
+	if secretRE == nil {
+		// we create a new regexp, as we cannot enforce, that nobody changed the SecretRE. This is unlike other
+		// functions, like Config.SetSecretRE, that have errors indicating a wrongful usage.
+		secretRE = regexp.MustCompile(SecretDefaultRE)
+	}
 
 	workQueue := make([]secretWorkItem, 1, initialWorkQueueDepth)
 	workQueue[0] = secretWorkItem{
@@ -52,7 +60,7 @@ func HideSecrets(node *yaml.Node, hideStructure bool) {
 		if currentWork.secret {
 			newWork = hideAll(currentWork.node, hideStructure)
 		} else {
-			newWork = hideSecrets(currentWork.node)
+			newWork = hideSecrets(currentWork.node, secretRE)
 		}
 
 		if len(newWork) > 0 {
@@ -68,21 +76,21 @@ func HideSecrets(node *yaml.Node, hideStructure bool) {
 // Mapping nodes are handled differently due to their structured key-value pair content.
 // Non-mapping nodes' content is scanned and added to the result with the secret flag set to false.
 // Returns nil if the provided YAML node is nil.
-func hideSecrets(node *yaml.Node) []secretWorkItem {
+func hideSecrets(node *yaml.Node, secretRE *regexp.Regexp) []secretWorkItem {
 	if node == nil {
 		return nil
 	}
 
 	var result []secretWorkItem
 
-	// just mapping nodes need special handling, is this step
+	// just mapping nodes need special handling in this step
 	if node.Kind == yaml.MappingNode {
 		result = make([]secretWorkItem, 0, len(node.Content)/2)
 
 		// The content is a sequence of key-value pairs, thus the content length is even.
 		// Subtracting 1 in the check accounts for potential uneven (invalid) content length.
 		for i := 0; i+1 < len(node.Content); i += 2 {
-			if SecretRE.MatchString(node.Content[i].Value) {
+			if secretRE.MatchString(node.Content[i].Value) {
 				result = append(result, secretWorkItem{
 					node:   node.Content[i+1],
 					secret: true,